requirements.txt
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
Automat==25.4.16
better-profanity==0.7.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
constantly==23.10.4
cryptography==45.0.6
cssselect==1.3.0
defusedxml==0.7.1
fastapi==0.116.1
filelock==3.19.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.12.1
itemloaders==1.3.2
jmespath==1.0.1
langdetect==1.0.9
lxml==6.0.1
packaging==25.0
parsel==1.10.0
profanity==1.1
Protego==0.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyDispatcher==2.0.7
pyOpenSSL==25.1.0
python-dotenv==1.1.1
queuelib==1.8.0
requests==2.32.5
requests-file==2.1.0
Scrapy==2.13.3
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
tldextract==5.3.0
Twisted==25.5.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
w3lib==2.3.1
zope.interface==7.2
